# Photo-prestiges/gateway/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Photo prestiges API docs",
    "description": "The documentation of all API calls for the Photo prestiges project made by Mayke van Beek and Charlotte Klokgieters",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user.",
        "description": "Register a new user using the given credentials.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the new user."
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email of the new user."
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "The password of the new user."
                  },
                  "role": {
                    "type": "string",
                    "enum": ["participant", "targetOwner"],
                    "description": "The role of the new user."
                  }
                },
                "required": ["username", "email", "password", "role"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful registration.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "The user has been successfully registered!"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error at the registration.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "Something went wrong while registering the new user."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/login": {
      "post": {
        "summary": "Log in.",
        "description": "Logging in with your username and password.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "description": "The username of the user."
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "description": "The password of the user."
                  }
                },
                "required": ["username", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully logged in.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "JWT-token for authenticated access."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Incorrect login credentials.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "The login credentials are incorrect."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/create-contest": {
      "post": {
        "summary": "Create a contest.",
        "description": "Creating a contest (without target).",
        "tags": ["contests"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "type": "string",
                    "description": "A description of the contest."
                  },
                  "endTime": {
                    "type": "string",
                    "format": "date-time",
                    "description": "The endtime of the contest."
                  }
                },
                "required": ["endTime"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contest created successfully. To add a target and place, use the /update-contest route.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "The contest has been created successfully."
                    },
                    "contestId": {
                      "type": "string",
                      "description": "The id of the contest that has been created. Needed when updating said contest."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error creating contest.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "Something went wrong while creating the contest."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/update-contest": {
      "put": {
        "summary": "Update an existing contest.",
        "description": "Updating an existing contest (with target).",
        "tags": ["contests"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "The id of the contest you'd like to update."
                  },
                  "place": {
                    "type": "string",
                    "description": "The place the target picture was taken."
                  },
                  "image": {
                    "type": "string",
                    "format": "base64",
                    "description": "The target picture."
                  }
                },
                "required": ["id", "place", "image"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contest updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "The contest has been updated successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid contest data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "Something went wrong while updating the contest."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/delete-contest": {
      "delete": {
        "summary": "Delete an existing contest.",
        "description": "Deleting an existing contest.",
        "tags": ["contests"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contestId": {
                    "type": "string",
                    "description": "The id of the contest you'd like to delete."
                  }
                },
                "required": ["contestId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contest deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "The contest has been deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid contest data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "Something went wrong while deleting the contest."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/register-for-contest": {
      "post": {
        "summary": "Register for a contest as a participant.",
        "description": "Here you can enter your submission for a contest.",
        "tags": ["contests"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contestId": {
                    "type": "string",
                    "description": "Id of the contest you want to register for."
                  }
                },
                "required": ["contestId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully entered the contest.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "You successfully entered the contest. To add a submission picture, use the /update-submission route."
                    },
                    "submissionId": {
                      "type": "string",
                      "description": "The id of the submission that has been created. Needed when updating said submission."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid contest data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Something went wrong with the contest data."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/update-submission": {
      "put": {
        "summary": "Update your submission.",
        "description": "Here you can update your submission for a contest by uploading a picture.",
        "tags": [
          "contests"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "submissionId": {
                    "type": "string",
                    "description": "Id of the submission you'd like to upload a picture to."
                  },
                  "image": {
                    "type": "string",
                    "format": "base64",
                    "description": "The target picture."
                  }
                },
                "required": ["submissionId", "image"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated your submission for the contest.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "You successfully updated your submission for the contest."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid submission data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Something went wrong with the submission data."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/delete-submission": {
      "delete": {
        "summary": "Delete an existing submission.",
        "description": "Deleting an existing submission.",
        "tags": ["contests"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "submissionId": {
                    "type": "string",
                    "description": "The id of the submission you'd like to delete."
                  }
                },
                "required": ["submissionId"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Submission deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "The submission has been deleted successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid submission data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "Something went wrong while deleting the submission."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/vote-for-contest": {
      "post": {
        "summary": "Vote for a contest.",
        "description": "Vote for a contest you're participating in.",
        "tags": ["contests"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "contestId": {
                    "type": "string",
                    "description": "Id of the contest you want to register for."
                  },
                  "thumbsUp": {
                    "type": "boolean",
                    "description": "Whether you'd like to give the contest thumbs up. If false, you'd give the target a thumbs down."
                  }
                },
                "required": ["contestId", "thumbsUp"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contest voted for successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "The contest has been voted for successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Error voting for contest.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "Something went wrong while voting for the contest."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string",
                      "description": "An internal server error occurred."
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "apiKey",
        "name": "authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}